[refactor] 린트 적용
