+++ sidebar/src/components/UI/organisms/bucket_contents/style.js
@@ -1,19 +1,19 @@
 import styled from 'styled-components';
 
 export const BucketContentsWrapper = styled.div`
-display: flex;
-flex-direction: column;
-justify-content: space-around;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
 `;
 
 export const titleStyle = {
-width: '100%',
-marginTop: '10px',
+  width: '100%',
+  marginTop: '10px',
 };
 
 export const decriptionStyle = {
-marginTop: '15px',
-width: '100%',
-rows: 4,
-rowsMax: 10,
+  marginTop: '15px',
+  width: '100%',
+  rows: 4,
+  rowsMax: 10,
 };
