+++ sidebar/src/components/UI/organisms/bucket_search/style.js
@@ -1,7 +1,9 @@
 import styled from 'styled-components';
 
-export const BucketSearchWrapper = styled.div`
-display: flex;
-padding: 15px;
-width: 500px;
+const BucketSearchWrapper = styled.div`
+  display: flex;
+  padding: 15px;
+  width: 500px;
 `;
+
+export default BucketSearchWrapper;
