+++ sidebar/src/components/UI/organisms/buttons/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import BucketCreateButton from '../../molecules/bucket_create_button';
-import { ButtonsWrapper } from './style.js'
+import ButtonsWrapper from './style';
 
 const Buttons = () => {
   return (
