+++ sidebar/src/components/UI/organisms/create_detail_list/style.js
@@ -1,5 +1,7 @@
 import styled from 'styled-components';
 
-export const DetailListWrapper = styled.div`
-margin-top: 15px;
+const DetailListWrapper = styled.div`
+  margin-top: 15px;
 `;
+
+export default DetailListWrapper;
